client-go/testing: support for finalizer
Set DeletionTimestamp if an object is deleted with finalizer present.
Delete object if DeletionTimestamp is present and finalizer is cleared.
Note that DeletionGracePeriodSeconds is still not supported, and is always treated as zero.

We should not use zero time as DeletionTimestamp, becasuse it is lost when going through json marshalling, which is required for patching.