apiserver/endpoints/installer: allow overriding hubGV
Strategic merge patch (k8s.io/apiserver/pkg/endpoints/handlers/patch.go)
assumes resources should be converted to an internal resource version,
presumably for etcd internal storage. It seems the example apiservice
gets around this by registering versions with the internal version but
this causes problems with much of the other tooling that presumes
resources have a version. At any rate, it seems reasonable to allow an
external apiservice to be able to choose its own hubGroupVersion while
still being able to utilize the heavy-lifting logic in
endpoints/installer.go and endpoints/handlers/patch.go.

To achieve this, allow the rest.Storage parameter to drive what the
hubGroupVersion is when it implements `rest.HubGroupVersionProvider`.

Note that a previous attempt tried to utilize the existing
`rest.StorageVersionProvider` but the versions for a storage version and
a hub version are not aligned closely enough to be unified. Namely, the
hub version uses `__internal` for all types, whereas a storage version
uses `__internal` for non-external types (e.g. `statefulsets/scale`) but
then uses the external group version elsewhere (e.g. `v1` for
`statefulsets`).

Signed-off-by: Andrew DeMaria <ademaria@cloudflare.com>