e2e: Test resizing memory requests only
This fix adds a test case where only memory requests are resized in
containers whose `resizePolicy` is `RestartContainer`. In this case
containers are not restarted because nothing is modified in the
runtime.

This fix also verifies resources in a pod status is updated earlier
than ever in test cases where only memory requests are resized.